log-publisher/requirements.txt
build==0.7.0
dataclasses-json==0.5.6
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
pep517==0.12.0
pika==1.2.0
pyparsing==3.0.6
python-dateutil==2.8.2
regex==2021.11.10
Rx==3.2.0
sh==1.14.2
six==1.16.0
tomli==2.0.0
typing-inspect==0.7.1
typing_extensions==4.0.1
ujson==5.1.0
